// client/src/Interfaces/serviceInterfaces.ts
/**
 * Интерфейс для глобальных уведомлений (Snackbar).
 */
export interface Notification {
   /** Флаг, указывающий, открыто ли уведомление */
   open: boolean;
   /** Текст сообщения уведомления */
   message: string;
   /** Тип уведомления (уровень важности) */
   severity: 'success' | 'error' | 'info';
}

/**
 * Интерфейс для настроек фильтрации задач.
 */
export interface Filter {
   /** Список выбранных приоритетов для фильтра */
   priorities: string[];
   /** Список выбранных статусов для фильтра */
   statuses: string[];
   /** Список выбранных названий досок для фильтра */
   boards: string[];
}